Change callcount default for t_startstop_valsf
The initial implementation of t_startstop_valsf used a default
callcount of 1 for a new timing event and a default callcount of
0 for any other calls. Upon reflection, this could be confusing.
This modified implementation sets the default callcount to 1
for any call to t_startstop_valsf. Also modified the underlying
GPTL commands to allow timing events with zero call counts and
to use more meaningful variable names.

[BFB]