clm4_5_17_r264
Changes from Dave Lawrence so that total vegetation carbon (totvegc) is used for resetting soil carbon
stocks during the AD spinup exit (to determine if you should reset both carbon and nitrogen). Also read
in totvegc for C13, and C14. These are then passed down into soilbiogechemistry nitrogenstate and carbonstate
restart to determine if the decomposition pools should be reset. The resetting is only done over non-crop
vegetation (ignoring both prognostic and generic crop types) and is only done on the exit_spinup phase
going out of accellerated decomposition mode.

Fix a couple small issues that in principle could change answers, but we didn't see answer changes in testing.
There was an error in how rootfr was initialized if it wasn't on the restart file. Made sure the crop patch
level restart variables are read after the enter_spinup code.