Add configuration for unit tests on skybridge
Add configuration needed to run unit tests on skybridge. This uses the
pFUnit library built on skybridge and installed in the ccsm project
space on the SNL capacity clusters. Configuration also added to the
compiler sections for ghost and chama, but PR #1914 needs to be applied
to the configuration for these machines before unit tests will run on
ghost and chama.