Rename and reorg upper level contents
Rename and reorg upper level contents.
Switch back to depth 2 at upper level because made it to hard to see all the content.
Expand the description of CIME in the top-most table of contents.
Reorganize the "what is cime" to put CCS first.
Remove multiple use of "CIME" from section headings.  We know its about CIME.
make part 1 and 2 about the Case Control System.
Move former part 3 on building a model with CIME to the end after data models and driver.