Reorganizing customizing and use case examples
Split up the "Customizing a case" material in to several files
and combine it with "Use cases" material.  Move it all to part 1 and
call it "Modifying a Case".  Move the pelayout material to part 2 in
the pes and threads section.  Move the mpi example to the Porting section in part 2.