Pass endrun msg to shr_sys_abort
I want to write some expected-exception tests for my init_interp
changes. For these to work robustly, I want to check the error message
returned against the expected error message (to make sure it's not
aborting for some completely different reason). In doing so, I noticed
that our endrun calls don't actually pass the abort message on to
shr_sys_abort. This needs to be done in order to assert against the
expected message.

Also: add a new optional argument to endrun, additional_msg, which is
written to iulog but not passed along to shr_sys_abort. This is to
support expected-exception unit testing, particularly thinking about the
common practice of writing out the line number: this is too volatile to
test against in unit tests. So a call could look like this:

  call endrun(msg='Informative message', additional_msg=errmsg(__FILE__, __LINE__))

Fixes #347