Add a new interpolation method, set_single_match
This is an alternative to set_mindist that can work when there is a
single matching input point for each output point (or possibly 0
matching inputs for an inactive output point). In this method, we only
look for points that share the same lat/lon as the output point.

This method is intended to be used for the case where we want to copy
areas from the input file to the output file (rather than maintaining
areas from the surface dataset); some of the logic here is set up the
way it is in order to support this copying of areas.