Delete the interp testmods and associated tests
Now that we're doing interpolation pretty much all the time, we are
covering the cases that these originally tested. Two examples are:

- Crop to non-crop: covered by
  ERP_Ld5.f10_f10_musgs.I1850Clm50Bgc.cheyenne_gnu.clm-default1850 (uses
  finidat
  clmi.I1850Clm50BgcCrop.1366-01-01.0.9x1.25_gx1v6_simyr1850_c180424.nc)

- Non-crop to crop: covered by
  SMS_Ld5.f10_f10_musgs.I1850Clm45BgcCrop.cheyenne_gnu.clm-crop1850 (uses
  finidat
  clmi.I1850Clm45BgcGs.0901-01-01.0.9x1.25_gx1v6_simyr1850_c180204.nc)

In addition, we cover the mindist logic via unit test. What's not
covered by unit test is determining if all variables are initialized
reasonably - particularly important for the non-crop to crop case. But
I'm not sure how we can really be confident of that even if we have a
system test covering this (other than the standard way of seeing that
baseline comparisons haven't changed... but the issue is: if new
behavior is introduced that changes answers for crops, then we could
expect baseline comparisons to change, and so could miss some
issue). That is to say: I'm not sure it's worth having system tests that
explicitly test these cases: it's probably enough to have the unit tests
plus incidental coverage via existing system tests.

Resolves #369