Revert "Point to new initial conditions file"
This reverts commit 23cdea6759897289cc4c76f83bfa9b7973e07a07.
This reverts commit 56aa1f24b15fe26e1b2e629ccfaf557c49d7be59.

There were answer changes for the new initial conditions file. One
problem with the new finidat file is that it doesn't have C isotope
information on it... that's a problem, but not the (only) problem with
these baseline failures, I think. Another difference is that two
variables have changed from double to int. I'm not sure if these
completely explain the problem, but I'm going to back out the change to
finidat, just using the new file for the test that needs it.