Add a section in the bug report for a brief summary
The main motivation is: If we change the title of a bug, then some
context can be lost for the original bug report. It could help if the
summary - something like the title - were repeated in the initial bug
report, so that the summary would remain intact even if the title is
changed.