Reduce modelio namelist creation time for large NINST
DART assimilation jobs consist of multiple cycles,
each of which needs new log file names in the modelio namelists.
When NINST is large, creating them from scratch is slow (and serial).
It's much more efficient to replace the file names in the existing
namelist files.  This code is effective, but something more general
may be worth coding.  For example, creating the modelio files
for the first instance, then making modified copies for the other
instances.